# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Para saber mais leia o README.md",
        "title": "Projeto do Desafio Go CRUD",
        "contact": {}
    },
    "host": "localhost:3429",
    "basePath": "/",
    "paths": {
        "/usuarios": {
            "get": {
                "description": "Lista todos os usuários.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios"
                ],
                "responses": {}
            }
        },
        "/usuarios/atualizar": {
            "put": {
                "description": "Passe um email de um usuario existente. Preencha somente os campos que quer atualizar.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios"
                ],
                "parameters": [
                    {
                        "description": "Corpo JSON",
                        "name": "JSON",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.usuarioUpdate"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/usuarios/criar": {
            "post": {
                "description": "Cria um usuário.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios"
                ],
                "parameters": [
                    {
                        "description": "Corpo JSON",
                        "name": "JSON",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.usuarioCreate"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/usuarios/deletar": {
            "delete": {
                "description": "Passe um ID de um usuário existente e sua senha para o deletar.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios"
                ],
                "parameters": [
                    {
                        "description": "Corpo JSON",
                        "name": "JSON",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.usuarioDelete"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "controllers.usuarioCreate": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "senha": {
                    "type": "string"
                }
            }
        },
        "controllers.usuarioDelete": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "senha": {
                    "type": "string"
                }
            }
        },
        "controllers.usuarioUpdate": {
            "type": "object",
            "properties": {
                "emailAtual": {
                    "type": "string"
                },
                "emailNovo": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "senhaAtual": {
                    "type": "string"
                },
                "senhaNova": {
                    "type": "string"
                }
            }
        }
    }
}